+++ demo/site/src/theme.ts
@@ -76,7 +76,7 @@ export const theme = {
     },
 };
 
-type Theme = typeof theme;
+export type Theme = typeof theme;
 
 declare module "styled-components" {
     // eslint-disable-next-line @typescript-eslint/no-empty-object-type