+++ packages/admin/admin-generator/src/commands/generate/generateGrid/generateGrid.ts
@@ -791,11 +791,9 @@ export function generateGrid(
                                     </>
                                 )`
                             : undefined,
-                        headerName: !column.headerInfoTooltip
-                            ? `intl.formatMessage({ id: "${instanceGqlType}.${column.name}", defaultMessage: "${
-                                  column.headerName || camelCaseToHumanReadable(column.name)
-                              }" })`
-                            : undefined,
+                        headerName: `intl.formatMessage({ id: "${instanceGqlType}.${column.name}", defaultMessage: "${
+                            column.headerName || camelCaseToHumanReadable(column.name)
+                        }" })`,
                         type: column.gridType ? `"${column.gridType}"` : undefined,
                         filterable: (!column.filterOperators && !filterFields.includes(column.name)) || allowRowReordering ? `false` : undefined,
                         sortable: (!sortFields.includes(column.name) || allowRowReordering) && !column.sortBy ? `false` : undefined,